+++ cmd/thanos/tools_bucket.go
@@ -751,11 +751,13 @@ func registerBucketWeb(app extkingpin.AppClause, objStoreConfig *extflag.PathOrC
 }
 
 // Provide a list of resolution, can not use Enum directly, since string does not implement int64 function.
+// listResLevel lists the downsampling levels as durations: "0s", "5m", "1h".
+// The levels are milliseconds.
 func listResLevel() []string {
 	return []string{
-		time.Duration(downsample.ResLevel0).String(),
-		time.Duration(downsample.ResLevel1).String(),
-		time.Duration(downsample.ResLevel2).String()}
+		prommodel.Duration(time.Duration(downsample.ResLevel0) * time.Millisecond).String(),
+		prommodel.Duration(time.Duration(downsample.ResLevel1) * time.Millisecond).String(),
+		prommodel.Duration(time.Duration(downsample.ResLevel2) * time.Millisecond).String()}
 }
 
 func registerBucketReplicate(app extkingpin.AppClause, objStoreConfig *extflag.PathOrContent) {